requirements.txt
pandas
fastapi
uvicorn[standard]
python-dotenv
elasticsearch[async]